LayoutTests/fast/media/ios/ipad/enable-MediaSource-API-in-modern-compatibility-mode.html

<!-- webkit-test-runner [ contentMode=desktop ] -->
<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<script src="../../../../resources/js-test.js"></script>
</head>
<body>
<script>
description("This test verifies that the MediaSource API is enabled when enabling modern compatibility mode.");
shouldBe("'MediaSource' in window", "true");
shouldBe("MediaSource.prototype", "Object.getPrototypeOf(new MediaSource)");
</script>
</body>
</html>
