LayoutTests/fast/media/ios/ipad/enable-MediaSource-API-in-modern-compatibility-mode-expected.txt

This test verifies that the MediaSource API is enabled when enabling modern compatibility mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'MediaSource' in window is true
PASS MediaSource.prototype is Object.getPrototypeOf(new MediaSource)
PASS successfullyParsed is true

TEST COMPLETE
